applied comments

+++ modules/mukhin_i_numerical_integration/src/integration_methods.cpp
@@ -1,8 +1,10 @@
 // Copyright 2024 Mukhin Ivan
 
 #include "include/integration_methods.h"
+#include <cstdint>
 #include <cstdlib>
 #include <stdexcept>
+#include <sys/types.h>
 
 double IntegrationMethods::integrate_trapezoid(
         const std::function<double(double)> &f,
@@ -13,7 +15,7 @@ double IntegrationMethods::integrate_trapezoid(
     }
     double step = std::abs(xn - x0) / static_cast<double>(n);
     double result = 0.0;
-    for (int i = 0; i < n; i++) {
+    for (uint32_t i = 0; i < static_cast<uint32_t>(n); i++) {
         double x_next = x0 + (i + 1) * step;
         result += (step) * (f(x_next) + f(x_next - step)) / 2.0;
     }
@@ -29,7 +31,7 @@ double IntegrationMethods::integrate_quads(
     }
     double step = std::abs(xn - x0) / static_cast<double>(n);
     double result = 0.0;
-    for (int i = 0; i < n; i++) {
+    for (uint32_t i = 0; i < static_cast<uint32_t>(n); i++) {
         double x_next = x0 + i * step;
         result += f(x_next + step / 2.0);
     }
